Fix `androidNavigationBar.enforceContrast` app config property not working
